recover from sound errors without killing plato.
The original scales program exited on any errors talking to the sound
device. This should not be a fatal condition in platoterm since sound
isn't critical and a user may well be doing other things with the sound
system.

This commit reworks those original assumptions via use of a flag and a
macro.

The flag is set by the FinishedSound routine which is always called on
an error condition. The macro is used to check the flag in various
places and cause a return.

For the routine that opens the devices and
does the initial setup we call the macro every time FinishedSound could
have just been called. this will cause the setup routine to end just as
it would have is calls to FinishedSound never returned like they did
previously.

For other routines that talk to the sound hardware we call the macro as
the first thing in the function that way if we run into an error
condition all the sound calls become no-ops.