Create preferences file if it does not exist.
If we don't get a file handle with MODE_OLDFILE create an empty file
with MODE_NEWFILE and then set the default preferences. This avoids
a bug I found where running PLATOTerm on a real floppy or hard disk
(instead of a mapped directory) would never create a preferences file.