remove debug printfs and add requester for fallback.
some systems do not have serial.device (or how I tested it) or may have
entered in the wrong unit number. in these cases instead of trying to
exit PLATOTerm we pop up the device requester and block io_init until it
finishes.

this commit also removes the debug printfs I was using so we don't
depend on -ldebug.